Common/Crypto: Reduce differences between 3.3.5 and 6.x branches
Code functionality shouldn't have been modified.

+++ src/common/Cryptography/ARC4.h
@@ -25,8 +25,8 @@
 class ARC4
 {
     public:
-        ARC4(uint8 len);
-        ARC4(uint8 *seed, uint8 len);
+        ARC4(uint32 len);
+        ARC4(uint8 *seed, uint32 len);
         ~ARC4();
         void Init(uint8 *seed);
         void UpdateData(int len, uint8 *data);
